Implented the SQL component into the lab 0 workflow and deleted testing page
