+++ client/src/components/exercise/lab0/pages/backend/SchemaDesignRepair.js
@@ -1,8 +1,39 @@
 import React from "react";
 import { Page } from "../../components/Page";
 import { ROUTES } from "../../../../../constants/lab0/index";
+import SQLQueryMock from "../../Mocks/SQLQueryMock";
 
 export const SchemaDesignRepair = () => {
+  const sqlTableData = {
+    query: "SELECT * FROM labs",
+    name: "labs",
+    columns: [
+      "id",
+      "labName",
+      "labShortName",
+      "category",
+      "shortDescription",
+      "about",
+      "reading",
+      "reinforcement",
+      "quiz",
+    ],
+    rows: [
+      {
+        id: 99,
+        labName: "Walkthrough Demo Lab",
+        labShortName: "Demo Lab",
+        category: "Accessibility",
+        shortDescription: "Learn about creating a lab from...",
+        about:
+          "In this demo lab, you will learn why it is important to create...",
+        reading: "(...)",
+        reinforcement: "(...)",
+        quiz: "(...)",
+      },
+    ],
+  };
+
   return (
     <Page nextPage={ROUTES.SECTION_SERVICE_CONTROLLER} completed>
       <Page.Header>
@@ -10,6 +41,14 @@ export const SchemaDesignRepair = () => {
           Database Design &amp; Schema Models Repair
         </Page.Header.Title>
       </Page.Header>
+      <div className="tw-m-4">
+        <SQLQueryMock
+          name={sqlTableData.name}
+          query={sqlTableData.query}
+          columns={sqlTableData.columns}
+          records={sqlTableData.rows}
+        />
+      </div>
     </Page>
   );
 };
